refactor(clipboard): consolidate copy helpers into modules/clipboard.ts
- Move copyToClipboard, initGlobalCopyToClipboardListener, initCopyContent
  into a new web_src/js/modules/clipboard.ts; delete features/clipboard.ts
  and features/copycontent.ts.
- Flip copyToClipboard signature to (content, target?) and make target
  optional — without it, copy runs silently (no feedback, no re-entrancy
  guard). Tighten target type to HTMLElement; drop redundant casts.
- Reassign the content parameter after awaiting the function form so TS
  narrows via control flow — removes the intermediate variable.
- Empty content now skips feedback entirely (no red X icon swap), per
  wxiaoguang's review.
- Route editor context-menu Cut/Copy through copyToClipboard so clippie
  is imported only by the clipboard module.
- Revert clippie to 4.1.15 (uses ClippieContent directly).
