+++ package.json
@@ -40,7 +40,7 @@
     "chart.js": "4.5.1",
     "chartjs-adapter-dayjs-4": "1.0.4",
     "chartjs-plugin-zoom": "2.2.0",
-    "clippie": "4.2.0",
+    "clippie": "4.1.15",
     "codemirror-lang-elixir": "4.0.1",
     "colord": "2.9.3",
     "compare-versions": "6.1.1",
